Raw version bump to ns-3.29

+++ src/core/examples/hash-example.cc
@@ -27,6 +27,8 @@
 
 /**
  * \file
+ * \ingroup core-examples
+ * \ingroup hash
  * Example usage of ns3::Hash.
  *
  * This example reads words from a list of files, creates a dictionary
@@ -357,7 +359,7 @@ public:
     //
     // Number of buckets = k = 2^bits
     long double k32 = 0xFFFFFFFF;
-    long double k64 = 0xFFFFFFFFFFFFFFFFULL;
+    long double k64 = static_cast<long double> (0xFFFFFFFFFFFFFFFFULL);
 
     long double n = m_nphrases;
     long double Ec32 = n * (n - 1) / ( 2 * k32) * (1 - (n - 2)/(3 * k32));
